omdlib: Move things to dedicated modules
More progress in splitting up the omdlib.main module into
more logical independent parts.
